feat : combined experiment script
- merged experiment scripts to single file
    - `run_Exp.py`
    - cfg.MODEL_TYPE must be set to determine the model
- edited default config files
