+++ compas/experiments/config/cnn.yaml
@@ -1,8 +1,7 @@
 MODEL_TYPE: cnn
-DATA_PATH: "data/custom/tmp_dataset_with_total_population2024.csv" # shape = 28 * 486
+DATA_PATH: "data/custom/tmp_dataset_with_total_population2024.csv"
 INPUT_STEPS: 3
 OUTPUT_STEPS: 1
-
 TEST_SIZE: 6
 VAL_SIZE: 6
 CNN:
