+++ compas/experiments/config/lstm.yaml
@@ -1,5 +1,5 @@
 MODEL_TYPE: lstm
-DATA_PATH: "data/custom/tmp_dataset_with_interest_col_modified.csv"
+DATA_PATH: "data/custom/tmp_dataset_with_total_population2024.csv"
 INPUT_STEPS: 3
 OUTPUT_STEPS: 1
 TEST_SIZE: 6
@@ -8,7 +8,7 @@ LSTM:
     NUM_LAYERS: 2
     HIDDEN_SIZE: 64
     DROPOUT: 0.0
-    BIDIRECTIONAL: True
+    BIDIRECTIONAL: False
 N_EPOCHS: 50
 BATCH_SIZE: 8
 MLFLOW_TRACKING_URI: "databricks"
